feat(github): totals on the list ops, and echo the query that actually ran
Takes up the two items the previous re-run left open, both raised by all
three evaluated models.

issue.list and pr.list now return totalCount. The objection that held this
back was that their rows come from GraphQL while a total comes from the
search index, so the count might disagree with the list it accompanies. That
was measured rather than assumed, and on every uncapped comparison available
the two agree exactly: 210/210 open issues, 66/66 open PRs, 33/33 labelled
Tooling, 143/143 with no:assignee, 141/141 on milestone 0.12.0. The mapping
needed one piece of care: `gh pr list --state closed` includes merged pull
requests (sampled 196 merged against 4 plain-closed) and search's `is:closed`
does too, so closed maps to is:closed rather than to unmerged-only. This
matters because issue.list is the op an agent reaches for when counting; it
filters richly and returns milestone, which search cannot, so "how many
issues are on milestone X" previously needed two operations.

All four paging ops now also return effectiveQuery, the query actually
executed. `-label:bug` against a repository with no `bug` label excludes
nothing and returns the unfiltered total, which is indistinguishable from the
filter having been dropped; every model flagged it and two proposed exactly
this remedy. The echo also exposes the type:issue discriminator gh appends
invisibly, which one model could previously only infer from a 403's URL.

The count request is still made only when the page is truncated, so the extra
load on GitHub's separately-throttled search endpoint is not paid when a
complete page already knows its own size.

The count machinery moves to github-broker-count.ts so all four ops share one
implementation and cannot drift into reporting counts differently. The
issue-3407 regressions move to their own test file, since p10b crossed the
800-line cap.
